Onset detection utilty, removed and rearranged some files.
