--- results/11_results_ode_to_joy.txt
@@ -1,3 +1,0 @@
-
-Solution(variables=[[7, 7, 7, 7, 7, 7, 7, 13, 7, 7, 13, 7, 7, 13, 13], [[4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 19, 20, 27, 34, 35], [4, 12, 19, 20, 27, 34, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35]]],objectives=[-0.4009131403567381],constraints=[])
-Solution(variables=[[7, 7, 7, 7, 7, 7, 7, 13, 7, 7, 13, 7, 7, 13, 13], [[4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 19, 20, 27, 34, 35], [4, 12, 19, 20, 27, 34, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35]]],objectives=[-0.4009131403567381],constraints=[])
