--- results/perason_results_elise.txt
@@ -1,31 +1,0 @@
-
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.033620506651623014],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[8, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03355261170753251],constraints=[])
-Solution(variables=[[7, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03341337829892052],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[8, 1, 4, 1, 4, 1, 1, 2, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03315042668902924],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 7, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.033413378298920514],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.033620506651623014],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 8, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03355261170753253],constraints=[])
-Solution(variables=[[4, 1, 8, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.0335526117075325],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 6, 1, 7, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03266793879682933],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.033620506651623014],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.033620506651623014],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
-Solution(variables=[[4, 1, 4, 1, 4, 1, 1, 2, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.03401720539041379],constraints=[])
