--- results/perason_results_elise_candidates.txt
@@ -1,31 +1,0 @@
-
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
-Solution(variables=[[27, 27, 27, 27, 27, 3, 27, 11, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.016101511818249942],constraints=[])
