--- results/perason_results_jingle.txt
@@ -1,31 +1,0 @@
-
-Solution(variables=[[10, 10, 10, 10, 10, 0, 7, 7, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.026058613408372077],constraints=[])
-Solution(variables=[[10, 7, 10, 10, 7, 10, 10, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02787721379722112],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 7, 10, 10, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02901025102542463],constraints=[])
-Solution(variables=[[10, 10, 10, 7, 10, 7, 10, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.027877213797221117],constraints=[])
-Solution(variables=[[7, 10, 10, 10, 10, 10, 7, 9, 2, 7, 7], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.026964093193966164],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 7, 10, 10, 10, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.027499732117762667],constraints=[])
-Solution(variables=[[10, 7, 10, 10, 10, 10, 10, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02901025102542463],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 6, 8, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02684303608434907],constraints=[])
-Solution(variables=[[10, 0, 10, 10, 10, 10, 7, 9, 5, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.025432664464879608],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 7, 8, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.027841001307681055],constraints=[])
-Solution(variables=[[10, 0, 10, 10, 10, 10, 10, 9, 2, 6, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.027405608304295705],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 10, 7, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02847793413647487],constraints=[])
-Solution(variables=[[7, 7, 10, 10, 10, 10, 0, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.026356529448289125],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 10, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.029620883335589522],constraints=[])
-Solution(variables=[[10, 10, 10, 7, 7, 10, 10, 9, 2, 6, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.027039564771018425],constraints=[])
-Solution(variables=[[10, 10, 10, 7, 10, 10, 10, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02901025102542463],constraints=[])
-Solution(variables=[[10, 7, 10, 10, 7, 7, 10, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02723534131569718],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 10, 9, 2, 6, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.028819858656203415],constraints=[])
-Solution(variables=[[10, 7, 10, 7, 10, 7, 10, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.026964093193966168],constraints=[])
-Solution(variables=[[10, 7, 10, 10, 10, 10, 10, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02876199436219436],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 7, 9, 5, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.026772204807588573],constraints=[])
-Solution(variables=[[10, 10, 7, 10, 10, 10, 7, 6, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.026406309683675633],constraints=[])
-Solution(variables=[[10, 10, 7, 10, 10, 10, 10, 10, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02778204583979528],constraints=[])
-Solution(variables=[[7, 10, 10, 10, 10, 10, 7, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.028136325498299475],constraints=[])
-Solution(variables=[[10, 6, 10, 10, 10, 10, 10, 8, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02652110154997695],constraints=[])
-Solution(variables=[[10, 10, 6, 10, 10, 10, 10, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02798731157548368],constraints=[])
-Solution(variables=[[10, 10, 10, 10, 10, 10, 10, 10, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.028390956181241885],constraints=[])
-Solution(variables=[[6, 10, 10, 10, 10, 7, 10, 9, 2, 12, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.025507193286436615],constraints=[])
-Solution(variables=[[10, 10, 10, 7, 10, 7, 10, 9, 2, 7, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.027877213797221117],constraints=[])
-Solution(variables=[[10, 10, 0, 10, 10, 10, 10, 9, 2, 1, 10], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.02845585530185273],constraints=[])
