--- results/perason_results_jingle_candidates.txt
@@ -1,31 +1,0 @@
-
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
-Solution(variables=[[28, 28, 28, 28, 28, 28, 28, 5, 28, 12, 28], [[28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [28, 35], [5, 13], [4, 12, 20, 28, 35], [4, 12, 20, 28, 35], [28, 35]]],objectives=[-0.006137381601274382],constraints=[])
