--- results/results_elise.txt
@@ -1,31 +1,0 @@
-
-Solution(variables=[[27, 2, 27, 27, 2, 8, 27, 2, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7890625],constraints=[])
-Solution(variables=[[27, 8, 27, 27, 8, 19, 27, 5, 2], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.765625],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 8], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[7, 2, 27, 27, 2, 8, 27, 7, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.78125],constraints=[])
-Solution(variables=[[4, 19, 27, 7, 27, 19, 8, 27, 11], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7578125],constraints=[])
-Solution(variables=[[3, 19, 27, 27, 7, 27, 19, 7, 6], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7421875],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[5, 19, 27, 27, 27, 19, 8, 7, 8], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7421875],constraints=[])
-Solution(variables=[[27, 8, 27, 27, 8, 19, 27, 5, 11], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.765625],constraints=[])
-Solution(variables=[[27, 8, 27, 27, 8, 27, 19, 11, 5], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.78125],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 8, 27, 27, 8, 19, 7, 27, 1], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7578125],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 5], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 2, 27, 27, 2, 8, 27, 27, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7890625],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 8, 27, 6, 27, 19, 7, 27, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7578125],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 6], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[2, 27, 8, 27, 8, 4, 27, 6, 1], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.796875],constraints=[])
-Solution(variables=[[27, 27, 8, 27, 8, 27, 19, 4, 11], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.796875],constraints=[])
-Solution(variables=[[27, 8, 27, 6, 27, 19, 4, 27, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7578125],constraints=[])
-Solution(variables=[[5, 27, 2, 27, 2, 19, 27, 6, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.828125],constraints=[])
-Solution(variables=[[5, 27, 8, 27, 8, 19, 27, 7, 7], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.8046875],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 1], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 27, 27, 8, 27, 34, 5, 27, 0], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 19, 27, 6, 27, 19, 27, 5, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.75],constraints=[])
-Solution(variables=[[27, 8, 27, 27, 8, 19, 7, 27, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7578125],constraints=[])
