--- results/results_elise_candidates.txt
@@ -1,31 +1,0 @@
-
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 3], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 34], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
-Solution(variables=[[27, 19, 27, 27, 27, 19, 27, 19, 19], [[27], [19, 27, 34], [27], [19, 27, 34], [27], [3, 11, 19, 27, 34], [19, 27, 34], [3, 11, 19, 27, 34], [3, 11, 19, 34]]],objectives=[-0.7265625],constraints=[])
