--- results_csv/fft_candidates_elise2.csv
@@ -1,31 +1,0 @@
-,runs,song,fitnesses,0,1,2,3,4,5,6,7,8,mse with target
-0,0,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-1,1,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-2,2,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-3,3,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-4,4,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-5,5,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-6,6,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-7,7,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-8,8,elise2.wav,87.0233122349132,27,34,27,34,27,11,34,11,19,349.6394540192648
-9,9,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-10,10,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-11,11,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-12,12,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-13,13,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-14,14,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-15,15,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-16,16,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-17,17,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-18,18,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-19,19,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-20,20,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-21,21,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-22,22,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-23,23,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-24,24,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-25,25,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-26,26,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-27,27,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-28,28,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
-29,29,elise2.wav,86.91842859216712,27,19,27,34,27,11,34,11,34,361.64637956603747
