--- results_csv/fft_candidates_jingle2.csv
@@ -1,31 +1,0 @@
-,runs,song,fitnesses,mse,0,1,2,3,4,5,6,7,8,9,10,mse with target
-0,0,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-1,1,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-2,2,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-3,3,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-4,4,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-5,5,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-6,6,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-7,7,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-8,8,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-9,9,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-10,10,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-11,11,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-12,12,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-13,13,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-14,14,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-15,15,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-16,16,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-17,17,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-18,18,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-19,19,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-20,20,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-21,21,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
-22,22,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-23,23,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-24,24,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-25,25,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-26,26,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-27,27,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-28,28,jingle2.wav,86.41565369882252,86.41565369882252,35,35,28,35,35,28,35,5,12,12,35,248.84024644527432
-29,29,jingle2.wav,86.41938391637649,86.41938391637649,28,35,35,28,35,35,35,5,12,12,35,248.84024644527432
