--- results_csv/fft_candidates_ode_to_joy.csv
@@ -1,31 +1,0 @@
-,runs,song,fitnesses,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,mse with target
-0,0,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-1,1,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-2,2,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-3,3,ode_to_joy.wav,73.92532211177588,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,19,305.44668912789945
-4,4,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-5,5,ode_to_joy.wav,74.34342391125304,20,20,20,28,28,20,20,4,12,12,4,20,4,4,4,19,277.5256014858969
-6,6,ode_to_joy.wav,74.30502285290424,4,20,20,28,28,20,4,4,12,12,4,4,20,12,4,18,319.1363396829511
-7,7,ode_to_joy.wav,74.18686644521085,20,20,20,28,28,20,4,4,12,12,12,4,20,4,4,18,305.44668912789945
-8,8,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-9,9,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-10,10,ode_to_joy.wav,73.92532211177588,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,19,305.44668912789945
-11,11,ode_to_joy.wav,74.20699624541166,20,20,20,28,28,20,4,4,12,12,12,4,20,4,4,19,305.44668912789945
-12,12,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-13,13,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-14,14,ode_to_joy.wav,74.41777351134115,4,20,20,28,28,20,4,12,12,12,4,4,20,4,4,18,319.1363396829511
-15,15,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-16,16,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-17,17,ode_to_joy.wav,74.20699624541166,20,20,20,28,28,20,4,4,12,12,12,4,20,4,4,19,305.44668912789945
-18,18,ode_to_joy.wav,74.30266709683077,4,20,20,28,28,20,4,4,12,12,4,4,20,12,4,19,319.1363396829511
-19,19,ode_to_joy.wav,73.92532211177588,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,19,305.44668912789945
-20,20,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-21,21,ode_to_joy.wav,74.20699624541166,20,20,20,28,28,20,4,4,12,12,12,4,20,4,4,19,305.44668912789945
-22,22,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-23,23,ode_to_joy.wav,73.92532211177588,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,19,305.44668912789945
-24,24,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-25,25,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-26,26,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-27,27,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-28,28,ode_to_joy.wav,73.88313276809538,20,20,20,28,28,20,4,4,12,12,4,20,4,12,4,18,305.44668912789945
-29,29,ode_to_joy.wav,74.30266709683077,4,20,20,28,28,20,4,4,12,12,4,4,20,12,4,19,319.1363396829511
