--- results_csv/fft_no_candidates_elise2.csv
@@ -1,31 +1,0 @@
-,runs,song,fitnesses,0,1,2,3,4,5,6,7,8,mse with target
-0,0,elise2.wav,57.15305642920248,7,7,7,7,7,11,7,34,2,478.4897556621024
-1,1,elise2.wav,57.132077544724126,7,34,7,7,7,11,7,2,7,602.7574906771486
-2,2,elise2.wav,57.32621827021141,7,7,7,7,7,11,34,2,7,602.7574906771486
-3,3,elise2.wav,57.132077544724126,7,34,7,7,7,11,7,2,7,602.7574906771486
-4,4,elise2.wav,57.15745801002164,7,7,7,34,7,2,7,11,7,560.1277554605273
-5,5,elise2.wav,57.15469692225096,7,7,7,2,7,11,7,7,34,599.4246180870517
-6,6,elise2.wav,57.0897942826708,7,2,7,7,7,7,7,11,34,571.1849452147495
-7,7,elise2.wav,58.09688837552545,7,7,7,34,7,2,7,2,11,558.3830000143982
-8,8,elise2.wav,57.10935921813187,7,34,7,2,7,7,7,11,7,609.1320605301427
-9,9,elise2.wav,57.13611246034751,7,34,7,7,2,11,7,7,7,512.6190085381481
-10,10,elise2.wav,57.14075166351621,2,7,7,34,7,11,7,7,7,512.6190085381481
-11,11,elise2.wav,57.17011652584138,7,7,7,7,7,2,34,11,7,560.1277554605273
-12,12,elise2.wav,57.131602105325854,7,7,7,2,7,7,34,11,7,609.1320605301427
-13,13,elise2.wav,57.03051737969417,7,2,7,7,7,34,7,11,7,563.7566555800325
-14,14,elise2.wav,57.04057017825332,7,34,7,7,2,7,7,7,11,478.2067441291233
-15,15,elise2.wav,57.11402451471531,7,34,7,7,7,11,7,7,2,561.3379983586022
-16,16,elise2.wav,57.030517379694174,7,7,7,34,7,11,7,7,2,561.3379983586022
-17,17,elise2.wav,57.030517379694174,7,7,7,34,7,11,7,7,2,561.3379983586022
-18,18,elise2.wav,57.10216815704328,7,34,7,7,7,7,7,2,11,568.3452262681237
-19,19,elise2.wav,57.18683549004709,2,7,7,7,7,34,7,7,11,432.8313391790131
-20,20,elise2.wav,57.08846642191268,7,7,7,7,7,2,34,7,11,553.9551639238045
-21,21,elise2.wav,57.131602105325854,7,7,7,2,7,7,34,11,7,609.1320605301427
-22,22,elise2.wav,57.10935921813187,7,34,7,2,7,7,7,11,7,609.1320605301427
-23,23,elise2.wav,57.107679526640794,7,2,7,7,7,7,34,11,7,609.1320605301427
-24,24,elise2.wav,57.06048042851398,7,7,2,34,7,7,7,7,11,478.2067441291233
-25,25,elise2.wav,57.241930626377126,7,7,7,34,2,7,7,7,11,478.2067441291233
-26,26,elise2.wav,57.151896279403765,7,2,7,7,7,34,7,7,11,557.5840640433098
-27,27,elise2.wav,57.08979428267081,7,7,7,7,7,11,34,7,2,561.3379983586022
-28,28,elise2.wav,57.08846642191267,7,7,7,7,2,34,7,11,7,439.0039307157358
-29,29,elise2.wav,57.08846642191267,7,7,7,7,2,34,7,11,7,439.0039307157358
