--- results_csv/fft_no_candidates_jingle2.csv
@@ -1,31 +1,0 @@
-,runs,song,fitnesses,mse,0,1,2,3,4,5,6,7,8,9,10,mse with target
-0,0,jingle2.wav,50.517226903556406,50.517226903556406,7,2,35,7,7,7,6,2,7,7,7,620.1003687254821
-1,1,jingle2.wav,50.46094981623323,50.46094981623323,7,7,6,7,7,6,35,10,7,7,7,659.2941911011109
-2,2,jingle2.wav,47.373187988161206,47.373187988161206,35,7,7,7,7,7,7,2,7,7,7,667.4730465206686
-3,3,jingle2.wav,49.824173145075335,49.824173145075335,7,35,7,10,7,7,2,7,2,7,7,724.3748159317681
-4,4,jingle2.wav,48.35083490653348,48.35083490653348,7,7,2,7,7,7,10,7,7,7,35,720.5502786056758
-5,5,jingle2.wav,48.36521385224426,48.36521385224426,7,35,7,7,10,7,7,2,7,7,7,655.619158410593
-6,6,jingle2.wav,48.33739079516515,48.33739079516515,7,7,7,10,7,35,7,2,7,7,7,655.619158410593
-7,7,jingle2.wav,47.29775933491953,47.29775933491953,7,7,35,7,10,7,7,7,7,7,7,761.9748489167953
-8,8,jingle2.wav,47.29775933491953,47.29775933491953,7,7,7,35,7,10,7,7,7,7,7,761.9748489167953
-9,9,jingle2.wav,49.596325996100816,49.596325996100816,7,2,7,7,35,7,7,10,10,7,7,612.8447524355909
-10,10,jingle2.wav,48.31194042921837,48.31194042921837,7,7,35,7,10,7,7,2,7,7,7,655.619158410593
-11,11,jingle2.wav,49.911500055789915,49.911500055789915,6,7,7,10,7,7,35,2,7,7,7,649.6710509265258
-12,12,jingle2.wav,48.34042168029612,48.34042168029612,7,7,7,2,7,7,35,7,10,7,7,715.4830833932168
-13,13,jingle2.wav,48.859792606462406,48.859792606462406,7,7,6,7,10,7,35,7,7,7,7,756.0267414327282
-14,14,jingle2.wav,48.85590460835005,48.85590460835005,7,7,7,35,7,7,10,6,7,7,7,731.3757465807486
-15,15,jingle2.wav,46.52386519666934,46.52386519666934,7,7,7,7,7,7,35,7,7,7,7,773.8287370268711
-16,16,jingle2.wav,49.81289363858974,49.81289363858974,10,2,35,7,7,2,7,7,7,7,7,679.1257082945563
-17,17,jingle2.wav,48.36043952745872,48.36043952745872,7,7,7,2,7,7,10,7,7,7,35,720.5502786056758
-18,18,jingle2.wav,47.3539303479504,47.3539303479504,7,35,7,2,7,7,7,7,7,7,7,732.4041667157516
-19,19,jingle2.wav,48.35254725767626,48.35254725767626,7,7,35,7,7,7,10,7,7,2,7,769.0471391379127
-20,20,jingle2.wav,47.372141648629096,47.372141648629096,7,7,7,35,7,7,7,7,7,2,7,780.9010272479885
-21,21,jingle2.wav,48.81075066508371,48.81075066508371,7,7,7,7,35,6,7,7,7,10,7,757.3941592941204
-22,22,jingle2.wav,48.32966538220484,48.32966538220484,7,2,35,7,7,7,7,7,10,7,7,715.4830833932168
-23,23,jingle2.wav,47.30431960105151,47.30431960105151,7,7,7,10,7,35,7,7,7,7,7,761.9748489167953
-24,24,jingle2.wav,49.864313366405234,49.864313366405234,2,35,10,7,7,7,7,6,7,7,7,689.9511762696291
-25,25,jingle2.wav,47.36616361703121,47.36616361703121,7,7,7,2,7,35,7,7,7,7,7,732.4041667157516
-26,26,jingle2.wav,48.77727212829361,48.77727212829361,7,2,2,35,7,7,7,7,7,7,7,690.979596404632
-27,27,jingle2.wav,47.29775933491953,47.29775933491953,7,35,7,10,7,7,7,7,7,7,7,761.9748489167953
-28,28,jingle2.wav,48.30164156988758,48.30164156988758,7,2,10,7,7,7,7,7,7,7,35,720.5502786056758
-29,29,jingle2.wav,47.304874878050924,47.304874878050924,7,7,7,7,10,35,7,7,7,7,7,761.9748489167953
