--- results_csv/fft_no_candidates_ode_to_joy.csv
@@ -1,31 +1,0 @@
-,runs,song,fitnesses,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,mse with target
-0,0,ode_to_joy.wav,56.95727663562501,7,7,20,7,15,7,15,15,15,6,15,15,15,7,10,5,464.6466243358818
-1,1,ode_to_joy.wav,57.96524577928615,6,10,20,7,5,7,9,15,15,1,15,7,14,7,7,29,394.51829294630903
-2,2,ode_to_joy.wav,57.080783612051036,15,14,5,15,7,7,20,10,7,7,9,7,7,2,7,29,357.4496220465639
-3,3,ode_to_joy.wav,57.93183831255192,5,15,7,7,15,15,15,15,7,15,6,4,20,7,2,7,380.91632326897326
-4,4,ode_to_joy.wav,58.82128953130259,20,5,15,7,28,15,6,7,7,2,9,15,7,4,7,18,387.2722387801823
-5,5,ode_to_joy.wav,57.776072273460095,7,5,7,28,15,7,15,7,15,15,2,15,1,15,15,19,407.43791196924536
-6,6,ode_to_joy.wav,56.776908138417625,20,5,7,15,15,7,15,7,10,6,7,15,7,7,15,7,396.048283011533
-7,7,ode_to_joy.wav,58.38913715929965,15,7,2,28,6,7,20,7,15,7,7,6,15,7,15,5,421.81700574489224
-8,8,ode_to_joy.wav,58.854911329321226,15,20,5,7,10,7,7,14,6,15,15,7,1,7,10,7,381.1383611939009
-9,9,ode_to_joy.wav,57.73716951767793,7,2,15,15,10,15,15,7,5,15,15,15,7,15,4,30,468.03911278130533
-10,10,ode_to_joy.wav,57.53573369466186,7,10,7,6,15,6,20,2,15,7,5,7,7,7,2,18,347.0792319986178
-11,11,ode_to_joy.wav,56.58767022803412,20,7,2,7,7,5,7,15,15,7,7,10,15,4,7,19,414.4971750680316
-12,12,ode_to_joy.wav,57.30634730784972,7,20,10,7,7,14,15,15,7,15,5,7,7,15,7,15,424.05732365772536
-13,13,ode_to_joy.wav,57.42306397434136,7,5,7,6,15,20,7,7,15,10,7,2,6,6,7,18,384.9961643125567
-14,14,ode_to_joy.wav,57.65724286102132,7,7,15,7,7,7,20,7,10,6,7,5,15,2,15,2,400.31753703943673
-15,15,ode_to_joy.wav,58.33481938307857,10,7,20,6,7,4,7,7,15,7,15,7,2,10,7,5,386.0062113947562
-16,16,ode_to_joy.wav,57.611285994350936,7,15,20,15,10,7,7,14,15,2,5,7,4,15,15,18,424.27946524988124
-17,17,ode_to_joy.wav,57.43641673651848,7,9,1,5,7,20,7,7,15,6,15,15,7,7,7,18,438.960442313507
-18,18,ode_to_joy.wav,57.31224266904914,7,5,20,7,28,15,7,2,6,15,15,15,7,7,10,18,398.62252602479373
-19,19,ode_to_joy.wav,57.533239777148765,10,1,15,7,15,7,2,15,15,15,7,20,5,7,7,7,333.4150965401063
-20,20,ode_to_joy.wav,57.6612286107612,7,15,9,5,15,2,7,15,7,15,7,4,15,10,15,19,422.71345082285865
-21,21,ode_to_joy.wav,57.951469505914304,2,5,6,7,7,10,10,7,15,15,10,15,20,15,7,23,352.1700315321152
-22,22,ode_to_joy.wav,57.80368369590078,15,15,5,15,7,15,7,15,7,10,15,1,7,15,6,15,446.78981037536175
-23,23,ode_to_joy.wav,58.5138187696533,7,20,7,5,15,7,2,6,4,15,15,10,7,7,4,38,379.3134076281802
-24,24,ode_to_joy.wav,57.4593494495968,15,7,15,7,6,15,15,2,5,15,10,7,15,7,15,15,478.26399723554056
-25,25,ode_to_joy.wav,58.36311793898914,2,14,7,15,2,20,7,7,10,15,7,7,15,15,7,5,406.1699408232906
-26,26,ode_to_joy.wav,57.9401617511992,15,7,10,4,2,7,7,6,15,15,7,20,7,5,15,15,395.8653103537644
-27,27,ode_to_joy.wav,58.141926968214385,7,20,7,7,15,15,10,7,6,2,7,20,15,7,15,5,371.10330871236664
-28,28,ode_to_joy.wav,58.1783771492902,10,7,20,7,15,8,15,7,7,15,15,10,5,7,7,26,373.03454278432895
-29,29,ode_to_joy.wav,57.94164156074899,20,7,7,15,5,10,7,7,15,7,6,10,15,14,15,23,367.84964364390305
