--- results_csv/fp_results_elise.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-elise,9,[27, 2, 27, 27, 2, 8, 27, 2, 3],-0.7890625,408.61990842728176
-elise,9,[27, 8, 27, 27, 8, 19, 27, 5, 2],-0.765625,344.2344319179877
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 8],-0.75,274.45013894288223
-elise,9,[7, 2, 27, 27, 2, 8, 27, 7, 19],-0.78125,501.14649475184655
-elise,9,[4, 19, 27, 7, 27, 19, 8, 27, 11],-0.7578125,302.14520611060715
-elise,9,[3, 19, 27, 27, 7, 27, 19, 7, 6],-0.7421875,464.79479648107355
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 34],-0.75,316.59601457100484
-elise,9,[5, 19, 27, 27, 27, 19, 8, 7, 8],-0.7421875,324.0643816681785
-elise,9,[27, 8, 27, 27, 8, 19, 27, 5, 11],-0.765625,337.02142456049415
-elise,9,[27, 8, 27, 27, 8, 27, 19, 11, 5],-0.78125,337.8723453270557
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 3],-0.75,275.7149151224569
-elise,9,[27, 8, 27, 27, 8, 19, 7, 27, 1],-0.7578125,355.13160698601934
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 5],-0.75,308.8173369040934
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 0],-0.75,325.0062776265964
-elise,9,[27, 2, 27, 27, 2, 8, 27, 27, 34],-0.7890625,461.51865290196247
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 3],-0.75,275.7149151224569
-elise,9,[27, 8, 27, 6, 27, 19, 7, 27, 0],-0.7578125,326.5328783493874
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 34],-0.75,316.59601457100484
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 34],-0.75,316.59601457100484
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 6],-0.75,327.2569003498769
-elise,9,[2, 27, 8, 27, 8, 4, 27, 6, 1],-0.796875,486.3613090867205
-elise,9,[27, 27, 8, 27, 8, 27, 19, 4, 11],-0.796875,390.4958317135329
-elise,9,[27, 8, 27, 6, 27, 19, 4, 27, 19],-0.7578125,319.1592803766576
-elise,9,[5, 27, 2, 27, 2, 19, 27, 6, 7],-0.828125,426.1577685629131
-elise,9,[5, 27, 8, 27, 8, 19, 27, 7, 7],-0.8046875,465.8367601111543
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 1],-0.75,288.231660259456
-elise,9,[27, 27, 27, 8, 27, 34, 5, 27, 0],-0.7265625,395.33866787611765
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 19],-0.75,288.0628986508862
-elise,9,[27, 19, 27, 6, 27, 19, 27, 5, 34],-0.75,316.59601457100484
-elise,9,[27, 8, 27, 27, 8, 19, 7, 27, 34],-0.7578125,383.49596129756816
