--- results_csv/fp_results_elise_candidates.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 3],-0.7265625,322.7435243977942
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 3],-0.7265625,322.7435243977942
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 3],-0.7265625,322.7435243977942
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 3],-0.7265625,322.7435243977942
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 34],-0.7265625,363.6246238463421
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
-elise,9,[27, 19, 27, 27, 27, 19, 27, 19, 19],-0.7265625,335.09150792622347
