--- results_csv/fp_results_jingle.csv
@@ -1,41 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 0, 28, 9],-0.76875,257.3312815861086
-jingle,12,[35, 6, 35, 35, 35, 35, 35, 6, 11, 9, 9, 6],-0.771875,399.59829538091816
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 13, 0, 6, 28, 8],-0.771875,343.55016669219077
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 10, 35, 11],-0.771875,296.4757206639415
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 11, 2, 10, 35, 8],-0.771875,366.8502833178225
-jingle,12,[4, 6, 35, 35, 35, 35, 35, 6, 11, 8, 28, 0],-0.759375,414.6158775784278
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 13, 9, 28, 10, 7],-0.775,365.2478883692438
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 0, 5, 2],-0.76875,284.8019955941428
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 6, 11, 8, 28, 4],-0.775,379.5468328483749
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 13, 0, 6, 35, 26],-0.771875,362.361262619718
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 6, 11, 9, 10, 7],-0.775,390.515962397069
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 6, 11, 9, 35, 25],-0.775,353.22498733610564
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 6, 11, 28, 28, 6],-0.7875,332.33814953997665
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 11, 2, 10, 35, 31],-0.771875,317.70542014678335
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 6, 28, 33],-0.771875,300.3534110052187
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 6, 11, 9, 0, 7],-0.775,370.1605074982035
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 13, 0, 10, 8, 24],-0.771875,310.9046673813423
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 13, 9, 28, 0, 10],-0.775,344.89243347037825
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 10, 10, 8],-0.771875,333.7666957249049
-jingle,12,[7, 35, 35, 35, 35, 6, 35, 11, 0, 6, 35, 6],-0.759375,421.8996312745179
-jingle,12,[7, 35, 35, 6, 35, 35, 35, 13, 0, 0, 35, 7],-0.75625,368.48399637164704
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 6, 35, 35],-0.771875,319.1645069327459
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 6, 11, 28, 28, 5],-0.7875,375.53490522694875
-jingle,12,[4, 4, 4, 4, 6, 5, 6, 7, 1, 8, 11, 9],-0.853125,620.8252323677359
-jingle,12,[4, 35, 35, 35, 35, 6, 35, 11, 0, 6, 35, 11],-0.759375,407.82381283353163
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 11, 2, 10, 1, 32],-0.76875,314.5782510867388
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 0, 6, 35, 21],-0.771875,319.1645069327459
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 13, 0, 6, 28, 12],-0.771875,349.49827417625784
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 11, 2, 35, 6, 8],-0.775,411.6130385100192
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 10, 12, 0, 22],-0.7625,287.50265058253916
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 6, 11, 28, 28, 8],-0.7875,375.53490522694875
-jingle,12,[35, 6, 35, 35, 35, 35, 35, 6, 11, 2, 8, 5],-0.771875,397.45330851602716
-jingle,12,[6, 6, 35, 35, 35, 35, 35, 11, 35, 28, 35, 12],-0.753125,380.4398906781957
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 6, 11, 8, 35, 4],-0.775,404.3060362599691
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 6, 11, 28, 28, 18],-0.7875,332.33814953997665
-jingle,12,[7, 35, 35, 35, 35, 35, 35, 11, 2, 10, 3, 38],-0.76875,370.4806801983814
-jingle,12,[7, 35, 35, 35, 35, 6, 35, 13, 0, 2, 28, 9],-0.75,387.56500406422657
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 11, 2, 11, 5, 1],-0.76875,315.68433918032207
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 6, 11, 28, 35, 25],-0.7875,394.34600115447597
-jingle,12,[6, 35, 35, 35, 35, 35, 35, 6, 11, 8, 9, 18],-0.775,401.53448113374253
