--- results_csv/fp_results_jingle_candidates.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 33],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 39],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 20],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 22],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 36],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 34],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 11],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 4, 35, 35, 6],-0.746875,283.98655860143487
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 15],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 21],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 29],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 34],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 5],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 38],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 22],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 14],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 4, 35, 35, 10],-0.746875,283.98655860143487
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 25],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 33],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 20],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 5],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 37],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 17],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 2],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 40],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 4],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 8],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 3],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 13],-0.75,303.86395443354104
-jingle,12,[35, 35, 35, 35, 35, 35, 35, 13, 35, 20, 35, 8],-0.75,303.86395443354104
