--- results_csv/fp_results_ode_to_joy.csv
@@ -1,33 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-ode_to_joy,16,[20, 35, 35, 28, 28, 35, 20, 4, 12, 12, 12, 3, 20, 4, 4, 10],-0.7321428571428571,331.19234590298475
-ode_to_joy,16,[4, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 20, 4, 4, 24],-0.7306547619047619,322.5995119001339
-ode_to_joy,16,[6, 4, 15, 28, 28, 35, 20, 12, 15, 12, 12, 4, 3, 13, 6, 29],-0.7619047619047619,416.83474579931755
-ode_to_joy,16,[0, 20, 15, 28, 28, 35, 20, 4, 12, 12, 7, 4, 4, 12, 4, 7],-0.7529761904761905,338.9145609260939
-ode_to_joy,16,[15, 35, 35, 28, 28, 4, 20, 6, 12, 6, 12, 5, 3, 4, 4, 22],-0.7470238095238095,381.9965055505162
-ode_to_joy,16,[0, 20, 15, 28, 28, 35, 20, 4, 12, 12, 7, 4, 3, 12, 12, 24],-0.7559523809523809,371.5643427929679
-ode_to_joy,16,[13, 5, 14, 28, 28, 20, 35, 12, 15, 12, 12, 6, 4, 4, 10, 13],-0.7619047619047619,378.6587102069025
-ode_to_joy,16,[4, 4, 15, 28, 28, 20, 35, 12, 15, 12, 13, 3, 35, 12, 3, 29],-0.7604166666666666,404.4374541056542
-ode_to_joy,16,[14, 4, 15, 28, 28, 35, 20, 4, 12, 12, 7, 4, 3, 4, 4, 25],-0.7529761904761905,338.20157660576217
-ode_to_joy,16,[15, 13, 14, 28, 28, 35, 20, 12, 6, 12, 12, 4, 1, 9, 4, 15],-0.7604166666666666,401.0021683728565
-ode_to_joy,16,[6, 35, 6, 28, 28, 20, 35, 12, 15, 12, 12, 4, 20, 12, 12, 30],-0.7529761904761905,380.48011317005745
-ode_to_joy,16,[15, 4, 15, 28, 28, 20, 35, 12, 15, 12, 12, 6, 4, 4, 14, 36],-0.7604166666666666,388.180752451298
-ode_to_joy,16,[35, 35, 6, 28, 28, 35, 20, 4, 12, 12, 4, 3, 20, 14, 4, 8],-0.7485119047619048,309.33180856118526
-ode_to_joy,16,[35, 35, 6, 28, 28, 20, 35, 12, 15, 12, 12, 15, 4, 12, 12, 25],-0.7559523809523809,412.41881128245
-ode_to_joy,16,[35, 35, 6, 28, 28, 35, 20, 4, 12, 12, 4, 20, 3, 13, 4, 7],-0.7455357142857143,367.0536693879314
-ode_to_joy,16,[4, 4, 15, 28, 28, 35, 20, 12, 6, 12, 12, 6, 3, 5, 15, 15],-0.7648809523809523,385.9305947907218
-ode_to_joy,16,[4, 5, 4, 28, 28, 35, 20, 6, 12, 14, 12, 3, 20, 4, 12, 0],-0.7529761904761905,351.5523519502397
-ode_to_joy,16,[14, 13, 14, 28, 28, 35, 20, 12, 6, 12, 13, 3, 35, 4, 4, 14],-0.7648809523809523,428.1312945445121
-ode_to_joy,16,[4, 35, 4, 28, 28, 35, 20, 6, 12, 12, 12, 3, 4, 13, 0, 0],-0.7529761904761905,381.94901732106604
-ode_to_joy,16,[35, 35, 6, 28, 28, 35, 20, 4, 12, 12, 7, 4, 3, 12, 4, 20],-0.7514880952380952,342.72365772536824
-ode_to_joy,16,[35, 35, 6, 28, 28, 35, 20, 12, 6, 12, 12, 4, 3, 12, 4, 35],-0.7619047619047619,383.39274329402616
-ode_to_joy,16,[35, 4, 15, 28, 28, 35, 20, 12, 6, 12, 12, 3, 13, 4, 11, 5],-0.7648809523809523,377.12574834780355
-ode_to_joy,16,[4, 20, 15, 28, 28, 35, 20, 12, 6, 12, 12, 4, 3, 15, 5, 10],-0.7678571428571429,359.6644680575353
-ode_to_joy,16,[12, 5, 15, 28, 28, 20, 35, 6, 12, 14, 12, 4, 3, 5, 12, 14],-0.7514880952380952,393.0592933350611
-ode_to_joy,16,[20, 35, 4, 28, 28, 35, 20, 12, 6, 12, 12, 7, 3, 5, 14, 13],-0.7633928571428571,346.81623687961644
-ode_to_joy,16,[13, 5, 15, 28, 28, 35, 20, 12, 6, 12, 12, 4, 3, 12, 4, 10],-0.7663690476190477,413.63461189581443
-ode_to_joy,16,[0, 20, 15, 28, 28, 20, 35, 4, 12, 12, 12, 3, 5, 12, 4, 12],-0.7514880952380952,398.04335449872576
-ode_to_joy,16,[0, 20, 15, 28, 28, 35, 20, 4, 12, 12, 5, 3, 4, 14, 12, 27],-0.7529761904761905,345.6086518940867
-ode_to_joy,16,[12, 11, 6, 28, 28, 35, 20, 12, 15, 12, 12, 4, 20, 12, 4, 0],-0.7559523809523809,378.95709040646193
-ode_to_joy,16,[14, 13, 14, 28, 28, 35, 20, 4, 12, 12, 7, 4, 3, 4, 12, 13],-0.7529761904761905,373.4873396397564
-ode_to_joy,16,[14, 4, 15, 28, 28, 20, 35, 4, 13, 6, 13, 4, 4, 12, 9, 9],-0.7485119047619048,393.62060386160425
-ode_to_joy,16,[35, 35, 6, 28, 28, 35, 20, 12, 6, 12, 12, 4, 3, 3, 12, 3],-0.7589285714285714,367.71317437691675
