--- results_csv/fp_results_ode_to_joy_candidates.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 24],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 20],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 13],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 22],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 8],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 6],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 30],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 25],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 24],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 35],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 13],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 25],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 16],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 21],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 12],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 16],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 29],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 22],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 4, 20, 4, 12, 12, 15],-0.7380952380952381,323.1412984320332
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 7],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 18],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 6],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 12],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 10],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 2],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 35],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 15],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 19],-0.7395833333333334,351.06238607403566
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 4, 3],-0.7395833333333334,336.8309489870848
-ode_to_joy,16,[20, 35, 20, 28, 28, 35, 20, 4, 12, 12, 12, 4, 4, 12, 12, 4],-0.7395833333333334,351.06238607403566
