--- results_csv/perason_results_elise.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 7],-0.033620506651623014,373.38477819532636
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[8, 1, 4, 1, 4, 1, 1, 2, 0],-0.03355261170753251,387.36075475501417
-elise,9,[7, 1, 4, 1, 4, 1, 1, 2, 0],-0.03341337829892052,455.5059320691691
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[8, 1, 4, 1, 4, 1, 1, 2, 7],-0.03315042668902924,380.35046722243817
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 7, 1, 1, 2, 0],-0.033413378298920514,455.5059320691691
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 7],-0.033620506651623014,373.38477819532636
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 8, 1, 1, 2, 0],-0.03355261170753253,387.36075475501417
-elise,9,[4, 1, 8, 1, 4, 1, 1, 2, 0],-0.0335526117075325,387.36075475501417
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 6, 1, 7, 1, 1, 2, 0],-0.03266793879682933,530.7938246008092
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 7],-0.033620506651623014,373.38477819532636
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 7],-0.033620506651623014,373.38477819532636
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
-elise,9,[4, 1, 4, 1, 4, 1, 1, 2, 0],-0.03401720539041379,380.3950657279023
