--- results_csv/perason_results_elise_candidates.csv
@@ -1,35 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-song,number of notes,result variables,objective value,mse with target
-song,number of notes,result variables,objective value,mse with target
-song,number of notes,result variables,objective value,mse with target
-song,number of notes,result variables,objective value,mse with target
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
-elise,9,[27, 27, 27, 27, 27, 3, 27, 11, 3],-0.016101511818249942,355.5909032007257
