--- results_csv/perason_results_jingle.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-jingle,12,[10, 10, 10, 10, 10, 0, 7, 7, 2, 7, 10],-0.026058613408372077,723.4654658546067
-jingle,12,[10, 7, 10, 10, 7, 10, 10, 9, 2, 7, 10],-0.02787721379722112,642.7362552569887
-jingle,12,[10, 10, 10, 10, 7, 10, 10, 9, 2, 1, 10],-0.02901025102542463,608.9661962373508
-jingle,12,[10, 10, 10, 7, 10, 7, 10, 9, 2, 7, 10],-0.027877213797221117,642.7362552569887
-jingle,12,[7, 10, 10, 10, 10, 10, 7, 9, 2, 7, 7],-0.026964093193966164,654.5901433670645
-jingle,12,[10, 10, 10, 10, 7, 10, 10, 10, 2, 7, 10],-0.027499732117762667,641.1825897958463
-jingle,12,[10, 7, 10, 10, 10, 10, 10, 9, 2, 1, 10],-0.02901025102542463,608.9661962373508
-jingle,12,[10, 10, 10, 10, 10, 10, 6, 8, 2, 1, 10],-0.02684303608434907,613.4380352704184
-jingle,12,[10, 0, 10, 10, 10, 10, 7, 9, 5, 1, 10],-0.025432664464879608,583.3614039841201
-jingle,12,[10, 10, 10, 10, 10, 10, 7, 8, 2, 1, 10],-0.027841001307681055,619.3861427544854
-jingle,12,[10, 0, 10, 10, 10, 10, 10, 9, 2, 6, 10],-0.027405608304295705,610.8753401580925
-jingle,12,[10, 10, 10, 10, 10, 10, 10, 7, 2, 7, 10],-0.02847793413647487,731.9670326433965
-jingle,12,[7, 7, 10, 10, 10, 10, 0, 9, 2, 7, 10],-0.026356529448289125,622.3808003581231
-jingle,12,[10, 10, 10, 10, 10, 10, 10, 9, 2, 7, 10],-0.029620883335589522,619.0284790368372
-jingle,12,[10, 10, 10, 7, 7, 10, 10, 9, 2, 6, 10],-0.027039564771018425,654.9385712771095
-jingle,12,[10, 10, 10, 7, 10, 10, 10, 9, 2, 1, 10],-0.02901025102542463,608.9661962373508
-jingle,12,[10, 7, 10, 10, 7, 7, 10, 9, 2, 1, 10],-0.02723534131569718,632.6739724575024
-jingle,12,[10, 10, 10, 10, 10, 10, 10, 9, 2, 6, 10],-0.028819858656203415,631.2307950569581
-jingle,12,[10, 7, 10, 7, 10, 7, 10, 9, 2, 7, 10],-0.026964093193966168,654.5901433670645
-jingle,12,[10, 7, 10, 10, 10, 10, 10, 9, 2, 7, 10],-0.02876199436219436,630.882367146913
-jingle,12,[10, 10, 10, 10, 10, 10, 7, 9, 5, 1, 10],-0.026772204807588573,603.7168588829857
-jingle,12,[10, 10, 7, 10, 10, 10, 7, 6, 2, 7, 10],-0.026406309683675633,725.0757065275012
-jingle,12,[10, 10, 7, 10, 10, 10, 10, 10, 2, 1, 10],-0.02778204583979528,619.2664188862841
-jingle,12,[7, 10, 10, 10, 10, 10, 7, 9, 2, 1, 10],-0.028136325498299475,620.8200843474266
-jingle,12,[10, 6, 10, 10, 10, 10, 10, 8, 2, 7, 10],-0.02652110154997695,635.3542061799805
-jingle,12,[10, 10, 6, 10, 10, 10, 10, 9, 2, 1, 10],-0.02798731157548368,603.0180887532838
-jingle,12,[10, 10, 10, 10, 10, 10, 10, 10, 2, 7, 10],-0.028390956181241885,629.3287016857705
-jingle,12,[6, 10, 10, 10, 10, 7, 10, 9, 2, 12, 10],-0.025507193286436615,610.3025139302838
-jingle,12,[10, 10, 10, 7, 10, 7, 10, 9, 2, 7, 10],-0.027877213797221117,642.7362552569887
-jingle,12,[10, 10, 0, 10, 10, 10, 10, 9, 2, 1, 10],-0.02845585530185273,576.7568532284096
