--- results_csv/perason_results_ode_to_joy.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-ode_to_joy,16,[10, 7, 2, 6, 9, 2, 7, 13, 14, 14, 13, 8, 7, 13, 13],-0.3573684020485436,616.6105999740832
-ode_to_joy,16,[2, 7, 2, 10, 1, 0, 7, 13, 0, 14, 13, 10, 6, 13, 13],-0.34390030588215625,585.6874519459203
-ode_to_joy,16,[6, 9, 7, 13, 6, 6, 7, 13, 2, 7, 13, 14, 1, 13, 13],-0.3436402056382101,565.5864411904454
-ode_to_joy,16,[2, 6, 0, 6, 14, 0, 7, 13, 7, 9, 13, 6, 7, 13, 13],-0.35148381035179244,653.123636128029
-ode_to_joy,16,[7, 9, 10, 7, 7, 7, 8, 13, 6, 6, 13, 0, 4, 13, 13],-0.3483810230662192,599.2012828819489
-ode_to_joy,16,[7, 1, 7, 7, 2, 14, 2, 13, 7, 2, 13, 9, 6, 13, 13],-0.35652053053933097,604.4356744849034
-ode_to_joy,16,[7, 7, 8, 7, 9, 3, 9, 13, 1, 12, 13, 2, 7, 13, 13],-0.3366132551166424,593.8978359466114
-ode_to_joy,16,[6, 6, 14, 2, 10, 6, 7, 13, 12, 2, 13, 6, 9, 13, 13],-0.3417198713294747,611.1612889292039
-ode_to_joy,16,[7, 2, 2, 7, 9, 10, 6, 13, 12, 9, 13, 6, 7, 13, 13],-0.3521113459558006,613.2002246123278
-ode_to_joy,16,[14, 1, 7, 14, 4, 2, 7, 13, 7, 7, 13, 2, 9, 13, 13],-0.3529237022752899,578.6367759492031
-ode_to_joy,16,[7, 7, 14, 7, 1, 7, 14, 13, 7, 10, 13, 9, 6, 13, 13],-0.36299270175987275,581.7508790117057
-ode_to_joy,16,[7, 8, 4, 14, 7, 4, 7, 13, 2, 6, 13, 4, 7, 13, 13],-0.3387485297256187,609.9319727009632
-ode_to_joy,16,[10, 1, 2, 6, 9, 0, 1, 13, 2, 2, 13, 7, 7, 13, 13],-0.34195943171374144,586.1746706405771
-ode_to_joy,16,[7, 10, 6, 10, 1, 0, 10, 13, 7, 7, 13, 10, 10, 13, 13],-0.34358739118279763,527.9065051185694
-ode_to_joy,16,[7, 6, 10, 2, 10, 14, 6, 13, 14, 6, 13, 7, 9, 13, 13],-0.35037376820821386,614.1616128892921
-ode_to_joy,16,[9, 10, 7, 14, 6, 7, 2, 13, 14, 7, 13, 7, 12, 13, 13],-0.35117582954621274,631.6020258304177
-ode_to_joy,16,[7, 12, 7, 9, 2, 14, 7, 13, 7, 6, 13, 7, 6, 13, 13],-0.3577185123094812,620.2036845060688
-ode_to_joy,16,[4, 9, 2, 14, 10, 6, 4, 13, 12, 6, 13, 3, 7, 13, 13],-0.3239308569528172,607.6970238866571
-ode_to_joy,16,[6, 7, 8, 7, 12, 7, 14, 13, 7, 14, 13, 14, 7, 13, 13],-0.35805292360454866,574.3274847738759
-ode_to_joy,16,[7, 7, 7, 2, 7, 6, 9, 13, 12, 9, 13, 7, 10, 13, 13],-0.35660120151227226,625.77960779232
-ode_to_joy,16,[7, 7, 7, 2, 6, 7, 4, 13, 9, 1, 13, 4, 1, 13, 13],-0.3435348391551126,612.6819359854866
-ode_to_joy,16,[7, 8, 3, 14, 7, 2, 6, 13, 0, 7, 13, 14, 7, 13, 13],-0.34786711107639773,624.195883547147
-ode_to_joy,16,[2, 7, 2, 14, 9, 14, 9, 13, 0, 12, 13, 0, 9, 13, 13],-0.33521420898149823,621.5324262450866
-ode_to_joy,16,[10, 8, 10, 6, 0, 12, 14, 13, 9, 14, 13, 7, 7, 13, 13],-0.3387818834013745,592.6034944494838
-ode_to_joy,16,[14, 6, 8, 14, 7, 1, 1, 13, 14, 12, 13, 9, 7, 13, 13],-0.341292161517619,607.2498293810203
-ode_to_joy,16,[1, 1, 7, 7, 7, 14, 0, 13, 8, 2, 13, 9, 7, 13, 13],-0.3489636692106883,595.2445164355752
-ode_to_joy,16,[14, 7, 6, 6, 7, 0, 8, 13, 12, 0, 13, 7, 7, 13, 13],-0.3488612123155112,603.6996587620405
-ode_to_joy,16,[6, 4, 7, 7, 2, 14, 6, 13, 9, 9, 13, 4, 7, 13, 13],-0.34440277905702665,632.138987516738
-ode_to_joy,16,[6, 6, 7, 8, 2, 14, 2, 13, 10, 12, 13, 1, 9, 13, 13],-0.3362470991871261,575.0245172994687
-ode_to_joy,16,[1, 9, 6, 10, 7, 0, 0, 13, 6, 7, 13, 7, 7, 13, 13],-0.35043674882933534,590.5194073690121
