--- results_csv/perason_results_ode_to_joy_candidates.csv
@@ -1,31 +1,0 @@
-song,number of notes,result variables,objective value,mse with target
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
-ode_to_joy,16,[4, 4, 20, 28, 28, 20, 4, 12, 12, 12, 12, 4, 4, 12, 12],-0.11239375170383673,389.20995205390693
