--- results_csv/test.csv
@@ -1,30 +1,0 @@
-620.1003687254821
-659.2941911011109
-667.4730465206686
-724.3748159317681
-720.5502786056758
-655.619158410593
-655.619158410593
-761.9748489167953
-761.9748489167953
-612.8447524355909
-655.619158410593
-649.6710509265258
-715.4830833932168
-756.0267414327282
-731.3757465807486
-773.8287370268711
-679.1257082945563
-720.5502786056758
-732.4041667157516
-769.0471391379127
-780.9010272479885
-757.3941592941204
-715.4830833932168
-761.9748489167953
-689.9511762696291
-732.4041667157516
-690.979596404632
-761.9748489167953
-720.5502786056758
-761.9748489167953
